requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==0.9.5
arrow==0.10.0
click==6.7
flask==0.12.2
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
lxml==4.0.0
mako==1.0.7               # via alembic
markupsafe==1.0           # via jinja2, mako
marshmallow==2.13.6
python-dateutil==2.6.1    # via alembic, arrow
python-editor==1.0.3      # via alembic
pyyaml==3.12
six==1.10.0               # via python-dateutil
sqlalchemy==1.1.14
werkzeug==0.12.2          # via flask
